requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Caching==2.3.0
Flask-Limiter==3.7.0
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
openai==1.40.0  # Ou grok-python se preferir
python-dotenv==1.0.1